Update to new curve functions

+++ pdfgen.c
@@ -1671,19 +1671,20 @@ int pdf_add_line(struct pdf_doc *pdf, struct pdf_object *page, float x1,
     return ret;
 }
 
-int pdf_add_cubic_bezier(struct pdf_doc *pdf, struct pdf_object *page, int x1,
-                         int y1, int x2, int y2, int xq1, int yq1, int xq2,
-                         int yq2, int width, uint32_t colour)
+int pdf_add_cubic_bezier(struct pdf_doc *pdf, struct pdf_object *page,
+                         float x1, float y1, float x2, float y2, float xq1,
+                         float yq1, float xq2, float yq2, float width,
+                         uint32_t colour)
 {
     int ret;
     struct dstr str = INIT_DSTR;
 
-    dstr_printf(&str, "%d w\r\n", width);
+    dstr_printf(&str, "%f w\r\n", width);
     dstr_printf(&str, "%d %d m\r\n", x1, y1);
     dstr_printf(&str, "/DeviceRGB CS\r\n");
     dstr_printf(&str, "%f %f %f RG\r\n", PDF_RGB_R(colour), PDF_RGB_G(colour),
                 PDF_RGB_B(colour));
-    dstr_printf(&str, "%d %d %d %d %d %d c S\r\n", xq1, yq1, xq2, yq2, x2,
+    dstr_printf(&str, "%f %f %f %f %f %f c S\r\n", xq1, yq1, xq2, yq2, x2,
                 y2);
 
     ret = pdf_add_stream(pdf, page, dstr_data(&str));
@@ -1693,20 +1694,21 @@ int pdf_add_cubic_bezier(struct pdf_doc *pdf, struct pdf_object *page, int x1,
 }
 
 int pdf_add_quadratic_bezier(struct pdf_doc *pdf, struct pdf_object *page,
-                             int x1, int y1, int x2, int y2, int xq1, int yq1,
-                             int width, uint32_t colour)
+                             float x1, float y1, float x2, float y2,
+                             float xq1, float yq1, float width,
+                             uint32_t colour)
 {
-    int xc1 = x1 + (xq1 - x1) * (2.0 / 3.0);
-    int yc1 = y1 + (yq1 - y1) * (2.0 / 3.0);
-    int xc2 = x2 + (xq1 - x2) * (2.0 / 3.0);
-    int yc2 = y2 + (yq1 - y2) * (2.0 / 3.0);
+    float xc1 = x1 + (xq1 - x1) * (2.0f / 3.0f);
+    float yc1 = y1 + (yq1 - y1) * (2.0f / 3.0f);
+    float xc2 = x2 + (xq1 - x2) * (2.0f / 3.0f);
+    float yc2 = y2 + (yq1 - y2) * (2.0f / 3.0f);
     return pdf_add_cubic_bezier(pdf, page, x1, y1, x2, y2, xc1, yc1, xc2, yc2,
                                 width, colour);
 }
 
 int pdf_add_custom_path(struct pdf_doc *pdf, struct pdf_object *page,
                         struct pdf_path_operation *operations,
-                        int operation_count, int stroke_width,
+                        int operation_count, float stroke_width,
                         uint32_t stroke_colour, uint32_t fill_colour)
 {
     int ret;
@@ -1717,7 +1719,7 @@ int pdf_add_custom_path(struct pdf_doc *pdf, struct pdf_object *page,
         dstr_printf(&str, "%f %f %f rg\r\n", PDF_RGB_R(fill_colour),
                     PDF_RGB_G(fill_colour), PDF_RGB_B(fill_colour));
     }
-    dstr_printf(&str, "%d w\r\n", stroke_width);
+    dstr_printf(&str, "%f w\r\n", stroke_width);
     dstr_printf(&str, "/DeviceRGB CS\r\n");
     dstr_printf(&str, "%f %f %f RG\r\n", PDF_RGB_R(stroke_colour),
                 PDF_RGB_G(stroke_colour), PDF_RGB_B(stroke_colour));
@@ -1726,22 +1728,22 @@ int pdf_add_custom_path(struct pdf_doc *pdf, struct pdf_object *page,
         struct pdf_path_operation operation = operations[i];
         switch (operation.op) {
         case 'm':
-            dstr_printf(&str, "%d %d m\r\n", operation.x1, operation.y1);
+            dstr_printf(&str, "%f %f m\r\n", operation.x1, operation.y1);
             break;
         case 'l':
-            dstr_printf(&str, "%d %d l\r\n", operation.x1, operation.y1);
+            dstr_printf(&str, "%f %f l\r\n", operation.x1, operation.y1);
             break;
         case 'c':
-            dstr_printf(&str, "%d %d %d %d %d %d c\r\n", operation.x1,
+            dstr_printf(&str, "%f %f %f %f %f %f c\r\n", operation.x1,
                         operation.y1, operation.x2, operation.y2,
                         operation.x3, operation.y3);
             break;
         case 'v':
-            dstr_printf(&str, "%d %d %d %d v\r\n", operation.x1, operation.y1,
+            dstr_printf(&str, "%f %f %f %f v\r\n", operation.x1, operation.y1,
                         operation.x2, operation.y2);
             break;
         case 'y':
-            dstr_printf(&str, "%d %d %d %d y\r\n", operation.x1, operation.y1,
+            dstr_printf(&str, "%f %f %f %f y\r\n", operation.x1, operation.y1,
                         operation.x2, operation.y2);
             break;
         case 'h':
